+++ src/component/Schema/Params/Page/assertDialog.js
@@ -2,7 +2,10 @@ import { buildAssertionTpl } from "service/assert";
 import { AssertDialog } from "../../Assert/AssertDialog";
 
 function renderBoolean( not ) {
-  return [ "true", "false" ].includes( not ) ? not : "false";
+  if ( typeof not !== "string" ) {
+    return "false";
+  }
+  return not;
 }
 
 export const assertDialog = {
@@ -15,11 +18,11 @@ export const assertDialog = {
   description: `Asserts there was called a dialog (alert, beforeunload, confirm or prompt).`,
   commonly: "assert dialog",
 
-  toLabel: ({ assert }) => `(${ renderBoolean( assert.not ) ? "none" : "any" }
+  toLabel: ({ assert }) => `(${ renderBoolean( assert.not ) === "true" ? "none" : "any" }
     of type \`${ assert.type }\` with \`${ assert.value }\`)`,
 
   toGherkin: ({ assert }) => `Assert that there were
-    called ${ renderBoolean( assert.not ) ? "no" : "any" } dialogs
+    called ${ renderBoolean( assert.not ) === "true" ? "no" : "any" } dialogs
     of type \`${ assert.type }\` with \`${ assert.value }\``,
 
   assert: {